--- Assets/MainMenu.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: a6a8ea1440495f144ade4ffb644cf9e0
