+++ Assets/Scenes/MainMenu.unity
@@ -2047,6 +2047,7 @@ GameObject:
   - component: {fileID: 1112760427}
   - component: {fileID: 1112760426}
   - component: {fileID: 1112760429}
+  - component: {fileID: 1112760430}
   m_Layer: 0
   m_Name: Main Camera
   m_TagString: MainCamera
@@ -2173,6 +2174,23 @@ MonoBehaviour:
     m_MipBias: 0
     m_VarianceClampScale: 0.9
     m_ContrastAdaptiveSharpening: 0
+--- !u!114 &1112760430
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1112760425}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: b53ac96a2f21c3f428fbad781d67f00d, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  input_field: {fileID: 0}
+  server_button_text: {fileID: 0}
+  server_text_color: {r: 0, g: 0, b: 0, a: 0}
+  tutorial_window: {fileID: 0}
+  credits_window: {fileID: 0}
 --- !u!1 &1214774441
 GameObject:
   m_ObjectHideFlags: 0
