+++ Assets/Tutorial/TutorialWall.prefab
@@ -11,7 +11,6 @@ GameObject:
   - component: {fileID: 8110581840618476112}
   - component: {fileID: 4998305883248392542}
   - component: {fileID: 9029948645385392475}
-  - component: {fileID: 3219251291073008462}
   m_Layer: 0
   m_Name: TutorialWall
   m_TagString: Untagged
@@ -135,22 +134,3 @@ BoxCollider2D:
   m_AutoTiling: 0
   m_Size: {x: 1, y: 1}
   m_EdgeRadius: 0
---- !u!114 &3219251291073008462
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 8204989269486249869}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: b78a59bb9acbf5c4981900c0675cf74b, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  tutorial_part_1: 0
-  tutorial_part_2: 0
-  tutorial_part_3: 0
-  tutorial_part_4: 0
-  tutorial_part_5: 0
-  tutorial_part_6: 0
-  tutorial_part_7: 0
